LayoutTests/fast/images/text-recognition/image-overlay-reversed-range-expected.txt


PASS textarea.value is "foo bar"
PASS successfullyParsed is true

TEST COMPLETE
